pyproject.toml
[tool.poetry]
name = "airflow-dbt"
version = "0.1.0"
description = ""
authors = ["alfredosa <asuarezaceves@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
apache-airflow = "^2.6.2"
apache-airflow-providers-postgres = "^5.5.0"
dbt-postgres = "^1.5.1"
pandas = "^2.0.2"
astronomer-cosmos = "^0.7.3"
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
